+++ internal/provider/data_source_variable_set.go
@@ -35,7 +35,7 @@ func dataSourceTFEVariableSet() *schema.Resource {
 				Computed: true,
 			},
 
-			"enforced": {
+			"priority": {
 				Type:     schema.TypeBool,
 				Computed: true,
 			},
@@ -93,7 +93,7 @@ func dataSourceTFEVariableSetRead(d *schema.ResourceData, meta interface{}) erro
 				d.Set("name", vs.Name)
 				d.Set("description", vs.Description)
 				d.Set("global", vs.Global)
-				d.Set("enforced", vs.Enforced)
+				d.Set("priority", vs.Priority)
 
 				// Only now include vars and workspaces to cut down on request load.
 				readOptions := tfe.VariableSetReadOptions{